Setting Up Project Folder
 	Introduction To Build Responsive Website
 	Important Note Before You Go Ahead
Important Note About Choosing Right Icons:
In next lecture, we are will be choosing the right icons for this  project and we are going to use Font-awesome library.
Recently Font-awesome has introduced new version of icons which is version 5 but this course is based on version 4.7 and URLs to Font-awesome library has been changed so therefore the http://fontawesome.io/ is replaced with https://fontawesome.com/v4.7.0/.
Please use https://fontawesome.com/v4.7.0/ for accessing Font-awesome library.

 	Setting Up Project Folder